Fix rdb part add partition size when larger than 4gb. Fix list command all option. Fix media-stream close and dispose issue. Add sector buffer to sector sector to support non sector dividable chunks. Initial support for .xz compressed images, write command can write .img.xz to physical drive.

+++ src/Hst.Imager.Core/Commands/RdbPartAddCommand.cs
@@ -155,7 +155,7 @@
             }
 
             var highCyl = lowCyl + cylinders - 1;
-            partitionSize = cylinders * cylinderSize;
+            partitionSize = (long)cylinders * cylinderSize;
 
             var partitionBlock = new PartitionBlock
             {
@@ -171,10 +171,6 @@
                 Sectors = (uint)((fileSystemBlockSize ?? 512) / rigidDiskBlock.BlockSize),
                 SizeBlock = rigidDiskBlock.BlockSize / SizeOf.Long
             };
-
-            // var partitionBlock =
-            //     PartitionBlock.Create(rigidDiskBlock, dosTypeBytes, name, partitionSize, fileSystemBlockSize ?? 512,
-            //         bootable);
 
             if (partitionBlock.HighCyl - partitionBlock.LowCyl + 1 <= 0)
             {
